rearranged where pytests fixtures are called to focus on prediction only
