+++ tests/test_predict.py
@@ -1,10 +1,49 @@
 import os
 from pathlib import Path
-
 import base64
+import shutil
 
-from tests import cfg_setup, cfg_teardown
+import pytest
+
 from math_notes import predict
+
+
+@pytest.fixture(scope="module")
+def cfg_setup():
+
+    # For now this is in configs.
+    path = Path("./tests_configs/")
+
+    # Setup.
+    if not os.path.isdir(path):
+        os.mkdir(path)
+
+    filename = "app_key2.txt"
+    filepath = path / filename
+    mode = "w+"
+    file_object = open(filepath, mode)
+    for i in range(1):
+        file_object.write("app_dummy_email_dot_com")
+    file_object.close()
+
+    filename = "app_id2.txt"
+    filepath = path / filename
+    mode = "w+"
+    file_object = open(filepath, mode)
+    for i in range(1):
+        file_object.write("ABCDEFGHIJKLMNOP1234567890")
+    file_object.close()
+
+
+@pytest.fixture(scope="module")
+def cfg_teardown():
+
+    # For now this is in configs.
+    path = Path("./tests_configs/")
+
+    # Teardown.
+    if os.path.isdir(path):
+        shutil.rmtree(path)
 
 
 def test_prediction_single(cfg_setup):
@@ -20,6 +59,8 @@ def test_prediction_single(cfg_setup):
     latex_return = predict._predict(image_uri)
 
     assert latex_return[0] == true_latex_string
+
+    # cfg_teardown()
 
 
 def test_prediction_multiple(cfg_setup):
@@ -42,7 +83,4 @@ def test_prediction_multiple(cfg_setup):
 
     assert latex_returns == true_latex_strings
 
-
-def test_call_teardown(cfg_teardown):
-    assert True == True
-    pass
+    # cfg_teardown()
